Rename file convertor.py to converter.py
