+++ kenja/convert.py
@@ -1,6 +1,6 @@
 from __future__ import absolute_import
 import argparse
-from kenja.convertor import HistorageConverter
+from kenja.converter import HistorageConverter
 
 
 class CommandParser:
